crud tweet authorized

+++ Pages/ShowTweet/DeleteTweet.cshtml.cs
@@ -2,6 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
@@ -11,15 +13,18 @@ using TwitterClone.Models;
 
 namespace TwitterClone.Pages.UserPortal
 {
+    [Authorize]
     public class DeleteTweet : PageModel
     {
         private readonly ILogger<DeleteTweet> _logger;
         private readonly TwitterCloneDbContext _context;
+        private readonly UserManager<User> _userManager;
 
-        public DeleteTweet(ILogger<DeleteTweet> logger,TwitterCloneDbContext context)
+        public DeleteTweet(ILogger<DeleteTweet> logger,TwitterCloneDbContext context,UserManager<User> userManager)
         {
             _logger = logger;
             _context=context;
+            _userManager=userManager;
         }
 
         [BindProperty]
@@ -33,8 +38,9 @@ namespace TwitterClone.Pages.UserPortal
 
             var tweet = await _context.Tweets.FirstOrDefaultAsync(t => t.Id == id);
             _logger.LogInformation("-----------------------------------id:" +tweet.Id);
+            var currentUser = await _userManager.GetUserAsync(HttpContext.User);
 
-            if (tweet == null)
+            if (tweet == null || tweet.Author.Id!=currentUser.Id)
             {
                 return NotFound();
             }
@@ -43,13 +49,14 @@ namespace TwitterClone.Pages.UserPortal
         }
         public async Task<IActionResult> OnPostAsync(int? id)
         {
-            if (id == null)
+            if (id == null )
             {
                 return NotFound();
             }
 
             var tweet = await _context.Tweets.FindAsync(id);
-            if (tweet != null)
+            var currentUser = await _userManager.GetUserAsync(HttpContext.User);
+            if (tweet != null && tweet.Author.Id==currentUser.Id)
             {
                 Tweet = tweet;
                 _context.Tweets.Remove(tweet);
